* Refactoring usecases

+++ src/main/java/ru/belov/ourabroad/api/usecases/delete/specialistservice/impl/DeleteSpecialistServiceUseCaseImpl.java
@@ -1,5 +1,6 @@
 package ru.belov.ourabroad.api.usecases.delete.specialistservice.impl;
 
+import io.micrometer.common.util.StringUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -17,13 +18,18 @@ public class DeleteSpecialistServiceUseCaseImpl
 
     @Override
     public void delete(String serviceId) {
-
+        if (serviceId == null || StringUtils.isBlank(serviceId)) {
+            log.error("Input id is null or empty");
+            throw new IllegalArgumentException();
+        }
+        log.info("[serviceId: {}] Start to delete service with ID: {}", serviceId, serviceId);
         boolean deleted = repository.deleteById(serviceId);
 
         if (!deleted) {
+            log.info("[serviceId: {}] Exception while deleting service", serviceId);
             throw new SpecialistServiceNotFoundException(serviceId);
         }
 
-        log.info("[serviceId={}] deleted", serviceId);
+        log.info("[serviceId: {}] deleted", serviceId);
     }
 }
